Create new Lightweight::Tools library to contain CxxModelPrinter and add some very initial tests to cover our current new use-case
