+++ src/tools/CMakeLists.txt
@@ -1,5 +1,5 @@
 find_package(yaml-cpp REQUIRED)
 add_executable(ddl2cpp ddl2cpp.cpp)
-target_link_libraries(ddl2cpp PRIVATE Lightweight::Lightweight yaml-cpp)
+target_link_libraries(ddl2cpp PRIVATE Lightweight::Lightweight Lightweight::Tools yaml-cpp)
 target_compile_features(ddl2cpp PUBLIC cxx_std_23)
 install(TARGETS ddl2cpp DESTINATION bin)
