--- src/tools/tests/1.sql
@@ -1,6 +1,0 @@
-CREATE TABLE "Person" (
-    "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-    "name" VARCHAR(25) NOT NULL,
-    "is_active" BOOLEAN NOT NULL,
-    "age" INTEGER
-);
