--- src/tools/tests/4.sql
@@ -1,18 +1,0 @@
--- --make-aliases
-CREATE TABLE IF NOT EXISTS User (
-    PK              INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-    NAME            VARCHAR(128) NOT NULL,
-    MAIL            VARCHAR(60) NOT NULL
-);
-CREATE TABLE IF NOT EXISTS TaskList (
-    PK              INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-    USER            INT NOT NULL,
-    CONSTRAINT      fk1 FOREIGN KEY (USER) REFERENCES user(PK)
-);
-CREATE TABLE IF NOT EXISTS TaskListEntry (
-    PK              INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-    TASKLIST        INT NOT NULL,
-    TIME            DATETIME NULL,
-    INFO            VARCHAR(255) NOT NULL,
-    CONSTRAINT      fk1 FOREIGN KEY (TASKLIST) REFERENCES TaskList(PK)
-);
